Move Loader and LoadResult to fixer internal
